+++ itest/babylon/babylon_e2e_test.go
@@ -9,10 +9,14 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/babylonlabs-io/finality-provider/eotsmanager/client"
+	eotscfg "github.com/babylonlabs-io/finality-provider/eotsmanager/config"
+	"github.com/babylonlabs-io/finality-provider/testutil"
 	"log"
 	"math/rand"
 	"os"
 	"path/filepath"
+	"strconv"
 	"testing"
 	"time"
 
@@ -508,4 +512,77 @@ func TestSigHeightOutdated(t *testing.T) {
 	require.Empty(t, res)
 
 	t.Logf("Finality signature for already finalized block at height %d was properly handled", finalizedBlock.Height)
+}
+
+func TestEotsdRollbackCmd(t *testing.T) {
+	t.Parallel()
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	testDir := t.TempDir()
+
+	eotsHomeDir := filepath.Join(testDir, "eots-home")
+	eotsCfg := eotscfg.DefaultConfigWithHomePath(eotsHomeDir)
+	eotsCfg.RPCListener = fmt.Sprintf("127.0.0.1:%d", testutil.AllocateUniquePort(t))
+	eotsCfg.Metrics.Port = testutil.AllocateUniquePort(t)
+
+	eh := e2eutils.NewEOTSServerHandler(t, eotsCfg, eotsHomeDir)
+	eh.Start(ctx)
+
+	eotsCli, err := client.NewEOTSManagerGRpcClient(eotsCfg.RPCListener, "")
+	require.NoError(t, err)
+
+	key, err := eh.CreateKey("eots-key-1")
+	require.NoError(t, err)
+
+	err = eotsCli.Ping()
+	require.NoError(t, err)
+
+	const numRecords = 100
+	const rollbackHeight = 10
+
+	for i := 0; i < numRecords; i++ {
+		_, err = eotsCli.SignEOTS(
+			key,
+			[]byte(testChainID),
+			[]byte("test"),
+			uint64(i),
+		)
+		require.NoError(t, err)
+	}
+
+	cmd := eotscmd.NewSignStoreRollbackCmd()
+	require.NoError(t, err)
+
+	err = eh.Stop()
+	require.NoError(t, err)
+
+	eotsPK, err := bbntypes.NewBIP340PubKey(key)
+	require.NoError(t, err)
+
+	cmd.SetArgs([]string{
+		"--home=" + eotsHomeDir,
+		"--rollback-until-height=" + strconv.Itoa(rollbackHeight),
+		"--chain-id=" + testChainID,
+		"--eots-pk=" + eotsPK.MarshalHex(),
+	})
+
+	err = cmd.Execute()
+	require.NoError(t, err)
+
+	eotsCfg.Metrics.Port = testutil.AllocateUniquePort(t)
+	eotsCfg.RPCListener = fmt.Sprintf("127.0.0.1:%d", testutil.AllocateUniquePort(t))
+	eh = e2eutils.NewEOTSServerHandler(t, eotsCfg, eotsHomeDir)
+	eh.Start(ctx)
+
+	for i := rollbackHeight; i <= numRecords; i++ {
+		exists, err := eh.IsRecordInDb(key, []byte(testChainID), uint64(i))
+		require.NoError(t, err)
+		require.False(t, exists)
+	}
+
+	for i := 0; i < rollbackHeight; i++ {
+		exists, err := eh.IsRecordInDb(key, []byte(testChainID), uint64(i))
+		require.NoError(t, err)
+		require.True(t, exists)
+	}
 }