+++ CHANGELOG.md
@@ -50,6 +50,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#382](https://github.com/babylonlabs-io/finality-provider/pull/382) chore: enforcing keyring-backend test in `fpd start` and `eotsd start`
 * [#397](https://github.com/babylonlabs-io/finality-provider/pull/397) chore: overwrite randomness proof
 * [#399](https://github.com/babylonlabs-io/finality-provider/pull/399) chore: submit immediately
+* [#410](https://github.com/babylonlabs-io/finality-provider/pull/410) chore: rollback sign records from eots store
 
 ## v1.0.0-rc.2